Split options into files
