+++ ext/src/options/options.html
@@ -136,11 +136,13 @@
     <div class="option">
       <h4 class="option-title">Quick Select Dub Languages</h4>
       <p class="option-blurb">
+        In the filter element (shown above) the options under "Enable Dubs for" let
+        you quickly pick the languages of the dubs you want to see or hide. </br></br>
         The "Shown Languages" box contains the languages available for quick switching on the filter UI element.
         The available languages are listed in the "Available Languages" box. Use the "Add" and "Remove" buttons
-        to move languages between the boxes. In the filter element (shown above) the options under "Enable Dubs for" let
-        you quickly pick the languages of the dubs you want to see or hide. Only 5 languages can be displayed as options
-        for Quick Select. Remember to save using the buttons at the bottom:
+        to move languages between the boxes. </br></br>
+        Only 5 languages can be displayed as options for Quick Select.
+        Remember to save using the button at the bottom:
       </p>
       <p id="lang-select-note" class="option-note"></p>
